requirements.txt
aiogram==3.7.0
gspread==5.12.0
google-auth==2.28.0
qrcode[pil]==7.4.2
pytz==2024.1
python-dotenv==1.0.1
apscheduler==3.10.4
Pillow==10.2.0
pyzbar~=0.1.9